Main test launcher now uses Python

+++ hdlcc/tests/test_persistency.py
@@ -34,24 +34,16 @@ import hdlcc.utils as utils
 _logger = logging.getLogger(__name__)
 
 CACHE_BUILD_SPEEDUP = 10
-BUILDER_NAME = os.environ.get('BUILDER_NAME', None)
-BUILDER_PATH = p.expandvars(os.environ.get('BUILDER_PATH', \
-                            p.expanduser("~/ghdl/bin/")))
 
 VIM_HDL_EXAMPLES_PATH = p.join(
     p.dirname(__file__), '..', '..', '.ci', 'test_support', 'vim-hdl-examples')
 
-if BUILDER_NAME is not None:
-    PROJECT_FILE = p.join(VIM_HDL_EXAMPLES_PATH, BUILDER_NAME + '.prj')
-else:
-    PROJECT_FILE = None
-
 class StandaloneProjectBuilder(hdlcc.HdlCodeCheckerBase):
     "Class for testing HdlCodeCheckerBase"
     _ui_handler = logging.getLogger('UI')
-    def __init__(self):
+    def __init__(self, *args, **kwargs):
         self._msg_queue = Queue()
-        super(StandaloneProjectBuilder, self).__init__(PROJECT_FILE)
+        super(StandaloneProjectBuilder, self).__init__(*args, **kwargs)
 
     def _handleUiInfo(self, message):
         self._ui_handler.info(message)
@@ -65,25 +57,38 @@ class StandaloneProjectBuilder(hdlcc.HdlCodeCheckerBase):
         self._ui_handler.error(message)
         self._msg_queue.put(('error', message))
 
-with such.A("hdlcc project using '%s' with persistency" % BUILDER_NAME) as it:
+with such.A("hdlcc project with persistency") as it:
 
     @it.has_setup
     def setup():
-        #  StandaloneProjectBuilder.cleanProjectCache(PROJECT_FILE)
-        utils.cleanProjectCache(PROJECT_FILE)
+        it.BUILDER_NAME = os.environ.get('BUILDER_NAME', None)
+        it.BUILDER_PATH = os.environ.get('BUILDER_PATH', None)
+        if not it.BUILDER_NAME:
+            return
+
+        it.PROJECT_FILE = p.join(VIM_HDL_EXAMPLES_PATH, it.BUILDER_NAME + '.prj')
+
+        #  StandaloneProjectBuilder.cleanProjectCache(it.PROJECT_FILE)
+        utils.cleanProjectCache(it.PROJECT_FILE)
 
         it.original_env = os.environ.copy()
         it.builder_env = os.environ.copy()
 
-        utils.addToPath(BUILDER_PATH)
+        #  utils.addToPath(it.BUILDER_PATH)
+        it.patch = mock.patch.dict(
+            'os.environ',
+            {'PATH' : os.pathsep.join([it.BUILDER_PATH, os.environ['PATH']])})
+        it.patch.start()
 
-        _logger.info("Builder name: %s", BUILDER_NAME)
-        _logger.info("Builder path: %s", BUILDER_PATH)
+        _logger.info("Builder name: %s", it.BUILDER_NAME)
+        _logger.info("Builder path: %s", it.BUILDER_PATH)
 
     @it.has_teardown
     def teardown():
-        #  StandaloneProjectBuilder.cleanProjectCache(PROJECT_FILE)
-        target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(PROJECT_FILE)
+        if not it.BUILDER_NAME:
+            return
+        #  StandaloneProjectBuilder.cleanProjectCache(it.PROJECT_FILE)
+        target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(it.PROJECT_FILE)
         if p.exists(target_dir):
             shutil.rmtree(target_dir)
 
@@ -96,12 +101,12 @@ with such.A("hdlcc project using '%s' with persistency" % BUILDER_NAME) as it:
 
     #      @it.has_teardown
     #      def teardown():
-    #          #  hdlcc.HdlCodeCheckerBase.cleanProjectCache(PROJECT_FILE)
-    #          utils.cleanProjectCache(PROJECT_FILE)
-    #          target_dir = hdlcc.config_parser.ConfigParser(PROJECT_FILE).getTargetDir()
+    #          #  hdlcc.HdlCodeCheckerBase.cleanProjectCache(it.PROJECT_FILE)
+    #          utils.cleanProjectCache(it.PROJECT_FILE)
+    #          target_dir = hdlcc.config_parser.ConfigParser(it.PROJECT_FILE).getTargetDir()
     #          if p.exists(target_dir):
     #              shutil.rmtree(target_dir)
-    #          cache_fname = utils.getDefaultCachePath(PROJECT_FILE)
+    #          cache_fname = utils.getDefaultCachePath(it.PROJECT_FILE)
     #          if p.exists(cache_fname):
     #              os.remove(cache_fname)
 
@@ -110,7 +115,7 @@ with such.A("hdlcc project using '%s' with persistency" % BUILDER_NAME) as it:
     #      def test_001():
     #          for _ in range(5):
     #              start = time.time()
-    #              project = StandaloneProjectBuilder()
+    #              project = StandaloneProjectBuilder(it.PROJECT_FILE)
     #              project.clean()
     #              parse_time = time.time() - start
     #              project.buildByDependency()
@@ -145,7 +150,7 @@ with such.A("hdlcc project using '%s' with persistency" % BUILDER_NAME) as it:
     #      def test_002():
     #          _logger.info("Creating object")
     #          start = time.time()
-    #          project = StandaloneProjectBuilder()
+    #          project = StandaloneProjectBuilder(it.PROJECT_FILE)
     #          parse_time = time.time() - start
     #          _logger.info("Building de facto")
     #          project.buildByDependency()
@@ -165,7 +170,7 @@ with such.A("hdlcc project using '%s' with persistency" % BUILDER_NAME) as it:
 
     @mock.patch('hdlcc.config_parser.hasVunit', lambda: False)
     def _buildWithoutCache():
-        it.project = StandaloneProjectBuilder()
+        it.project = StandaloneProjectBuilder(it.PROJECT_FILE)
         it.project.waitForBuild()
         it.project._saveCache()
 
@@ -185,7 +190,7 @@ with such.A("hdlcc project using '%s' with persistency" % BUILDER_NAME) as it:
     def _buildWithCache():
         del it.project
 
-        it.project = StandaloneProjectBuilder()
+        it.project = StandaloneProjectBuilder(it.PROJECT_FILE)
         it.project.waitForBuild()
 
         messages = []
@@ -203,35 +208,49 @@ with such.A("hdlcc project using '%s' with persistency" % BUILDER_NAME) as it:
     with it.having('an undecodable cache file'):
         @it.has_setup
         def setup():
-            #  cache_fname = utils.getDefaultCachePath(PROJECT_FILE)
-            target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(PROJECT_FILE)
+            if not it.BUILDER_NAME:
+                return
+            #  cache_fname = utils.getDefaultCachePath(it.PROJECT_FILE)
+            target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(it.PROJECT_FILE)
             if p.exists(target_dir):
-                _logger.fatal("Target dir '%s' removed", target_dir)
+                _logger.info("Target dir '%s' removed", target_dir)
                 shutil.rmtree(target_dir)
             else:
-                _logger.fatal("target_dir '%s' not found", target_dir)
+                _logger.info("target_dir '%s' not found", target_dir)
 
         @it.has_teardown
         def teardown():
-            #  hdlcc.HdlCodeCheckerBase.cleanProjectCache(PROJECT_FILE)
-            target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(PROJECT_FILE)
+            if not it.BUILDER_NAME:
+                return
+            #  hdlcc.HdlCodeCheckerBase.cleanProjectCache(it.PROJECT_FILE)
+            target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(it.PROJECT_FILE)
             shutil.rmtree(target_dir)
 
         @it.should('build a project without cache')
         def test_001():
+            if not it.BUILDER_NAME:
+                _logger.info("Skipping test, it requires a builder")
+                return
+
             _buildWithoutCache()
 
         @it.should('recover from cache')
         def test_002():
+            if not it.BUILDER_NAME:
+                _logger.info("Skipping test, it requires a builder")
+                return
             _buildWithCache()
 
         @it.should('build without cache if cache is invalid')
         @mock.patch('hdlcc.config_parser.hasVunit', lambda: False)
         def test_003():
-            target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(PROJECT_FILE)
+            if not it.BUILDER_NAME:
+                _logger.info("Skipping test, it requires a builder")
+                return
+            target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(it.PROJECT_FILE)
             cache_fname = p.join(target_dir, '.hdlcc.cache')
             open(cache_fname, 'w').write("hello")
-            project = StandaloneProjectBuilder()
+            project = StandaloneProjectBuilder(it.PROJECT_FILE)
             project.waitForBuild()
             project._saveCache()
 
@@ -251,45 +270,60 @@ with such.A("hdlcc project using '%s' with persistency" % BUILDER_NAME) as it:
     with it.having('the builder working folder erased'):
         @it.has_setup
         def setup():
-            target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(PROJECT_FILE)
+            if not it.BUILDER_NAME:
+                return
+            target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(it.PROJECT_FILE)
             if p.exists(target_dir):
                 shutil.rmtree(target_dir)
 
         @it.has_teardown
         def teardown():
-            target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(PROJECT_FILE)
+            if not it.BUILDER_NAME:
+                return
+            target_dir, _ = hdlcc.config_parser.ConfigParser.simpleParse(it.PROJECT_FILE)
             if p.exists(target_dir):
                 shutil.rmtree(target_dir)
 
         @it.should('build without cache if cache is invalid')
         def test_001():
+            if not it.BUILDER_NAME:
+                _logger.info("Skipping test, it requires a builder")
+                return
             _buildWithoutCache()
-            target_dir = hdlcc.config_parser.ConfigParser(PROJECT_FILE).getTargetDir()
+            target_dir = hdlcc.config_parser.ConfigParser(it.PROJECT_FILE).getTargetDir()
             shutil.rmtree(target_dir)
             _buildWithoutCache()
 
     with it.having('the builder failing to run'):
         @it.has_setup
         def setup():
-            cache_fname = utils.getDefaultCachePath(PROJECT_FILE)
+            if not it.BUILDER_NAME:
+                return
+            cache_fname = utils.getDefaultCachePath(it.PROJECT_FILE)
             if p.exists(cache_fname):
                 os.remove(cache_fname)
 
         @it.has_teardown
         def teardown():
-            utils.cleanProjectCache(PROJECT_FILE)
+            if not it.BUILDER_NAME:
+                return
+            utils.cleanProjectCache(it.PROJECT_FILE)
 
         @it.should('use fallback builder if recovering cache failed')
         @mock.patch('hdlcc.config_parser.hasVunit', lambda: False)
         def test_001():
+            if not it.BUILDER_NAME:
+                _logger.info("Skipping test, it requires a builder")
+                return
             _logger.info("Building without cache")
             _buildWithoutCache()
             _logger.info("Restoring original path")
-            utils.removeFromPath(BUILDER_PATH)
+            #  utils.removeFromPath(it.BUILDER_PATH)
+            it.patch.stop()
 
             _logger.info("Building with changed env")
 
-            project = StandaloneProjectBuilder()
+            project = StandaloneProjectBuilder(it.PROJECT_FILE)
             time.sleep(1)
             project.waitForBuild()
             time.sleep(1)
@@ -321,5 +355,4 @@ with such.A("hdlcc project using '%s' with persistency" % BUILDER_NAME) as it:
             return matches, messages
 
 it.createTests(globals())
-
